Fix error when splitting stat string only by space
observe:
me@labo:/proc$ cat /proc/79/stat
79 (DWC Notificatio) I 2 0 0 0 -1 69238880 0 0 0 0 0 0 0 0 0 -20 1 0 \
156 0 0 4294967295 0 0 0 0 0 0 0 2147483647 0 0 0 0 17 2 0 0 0 0 0 0 \
0 0 0 0 0 0 0

When splitting simply by space, the 2nd field is "DWC" and the 3rd is
"Notificatio".  This is not what we want. We want the 3rd field to be
an integer for the parent process ID. Otherwise the
strconv.Atoi(stat[3]) fails with a parse error on I. Hence we need an
exception on the splits if the space is encapsulated by parentheses.